feat(schema): Add fields which are sensitive in nature and Nix development setup
This adds fields to the metadata that the shop needs which are currently
stored in the manifest. The reason to move them here is that only the
merchant can update the manifest while currently the relay itself can
update the manifest along with the merchant, therefor moving them here
will make the system more secure.

- Add required fields to JSON schema including payment addresses, accepted currencies, and pricing currency.
- Implement Nix flake for development environment management.
- Update README with schema documentation and setup instructions.